pyproject.toml
[project]
name = "videodub"
version = "0.2.0"
description = "Professional video subtitle generation tool using OpenAI Whisper"
readme = "README.md"
authors = [{name = "Dragon", email = "fldx123456@163.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["video", "subtitle", "whisper", "speech-recognition", "translation"]
dependencies = [
    "openai-whisper>=1.0.0",
    "torch>=1.9.0",
    "torchaudio>=0.12.0",
    "transformers>=4.21.0",
    "psutil>=5.8.0",
    "ffmpeg-python>=0.2.0",
    "numpy>=1.21.0",
    "sentencepiece>=0.1.99",
    "accelerate>=0.20.3",
    "tqdm>=4.64.0",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
]

[project.urls]
Homepage = "https://github.com/Dragon/VideoDub"
"Bug Reports" = "https://github.com/Dragon/VideoDub/issues"
"Source" = "https://github.com/Dragon/VideoDub"

[project.scripts]
videodub = "videodub.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=videodub --cov-report=html --cov-report=term-missing"
testpaths = [
    "tests",
]

[tool.coverage.run]
source = ["src/videodub"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
]

[tool.mypy]
python_version = "3.9"
# warn_return_any = true
warn_unused_configs = true
# disallow_untyped_defs = true
disallow_incomplete_defs = true
# check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "ffmpeg",
    "whisper",
    "transformers.*",
    "psutil",
    "torch.*",
    "torchaudio.*",
]
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"